Fix all existing tests
